Moving all auto approve checks to permissions to fail rule creation instead of approval
